+++ src/Oscillator.h
@@ -8,6 +8,7 @@
 #ifndef Oscillator_h
 #define Oscillator_h
 
+#include <stdint.h>
 #include <Servo.h>
 
 //-- Macro for converting from degrees to radians
@@ -22,13 +23,13 @@ class Oscillator
     void attach(int pin, bool rev =false);
     void detach();
     
-    void SetA(unsigned int A) {_A=A;};
-    void SetO(unsigned int O) {_O=O;};
+    void SetA(uint16_t A) {_A=A;};
+    void SetO(uint16_t O) {_O=O;};
     void SetPh(double Ph) {_phase0=Ph;};
-    void SetT(unsigned int T);
-    void SetTrim(int trim){_trim=trim;};
-    int getTrim() {return _trim;};
-    void SetPosition(int position); 
+    void SetT(uint16_t T);
+    void SetTrim(int8_t trim){_trim=trim;};
+    int8_t getTrim() {return _trim;};
+    void SetPosition(uint8_t position); 
     void Stop() {_stop=true;};
     void Play() {_stop=false;};
     void Reset() {_phase=0;};
@@ -42,21 +43,21 @@ class Oscillator
     Servo _servo;
     
     //-- Oscillators parameters
-    unsigned int _A;  //-- Amplitude (degrees)
-    unsigned int _O;  //-- Offset (degrees)
-    unsigned int _T;  //-- Period (miliseconds)
+    uint16_t _A;  //-- Amplitude (degrees)
+    uint16_t _O;  //-- Offset (degrees)
+    uint16_t _T;  //-- Period (miliseconds)
     double _phase0;   //-- Phase (radians)
     
     //-- Internal variables
-    int _pos;         //-- Current servo pos
-    int _trim;        //-- Calibration offset
+    int16_t _pos;         //-- Current servo pos
+    int8_t _trim;        //-- Calibration offset
     double _phase;    //-- Current phase
     double _inc;      //-- Increment of phase
     double _N;        //-- Number of samples
-    unsigned int _TS; //-- sampling period (ms)
+    uint16_t _TS; //-- sampling period (ms)
     
-    long _previousMillis; 
-    long _currentMillis;
+    uint32_t _previousMillis; 
+    uint32_t _currentMillis;
     
     //-- Oscillation mode. If true, the servo is stopped
     bool _stop;
